Restructured "Get single service" response schema.
Extracted shared classes into shared classes file. Moved their tests to a separate tests file.
